Implementing the Flags Register
Had a lot of trouble implementing this, especially since I wanted to capture the challenge flags. Had a hard time understanding the AF flag but in the end I got fed up and made a truth table on all the possible bit combinations between 4-th(3) and 5-th(4) byte and solved it this way... Parity Flag I got a solution on the count from someone in the comments, but the homework results had the first command have the Parity Flag on, but I didn't understood why because it didn't made sense based on Intel's explanation(instruction resulted in 7 ,1bits... but parity should happen on even no odd bits). However testing the asm on emu8086 had that flag turned on as well...so I have no idea why. Also the Carry flag is turned on on the last challenge instruction but the result is 0....so why there is a carry flag there I don't know. Anyway moving to implementing jumps